+++ HomeWork/src/HomeWork3/task6/ICalculator.java
@@ -1,11 +1,11 @@
 package HomeWork3.task6;
 
 public interface ICalculator {
-    public float div(float a, float b);
-    public float mult (float a, float b);
-    public float sub(float a, float b);
-    public float add (float a, float b);
-    public float pow (float number, int power);
-    public float abs (float a);
-    public float sqrt (float a);
+    public double div(double a, double b);
+    public double mult (double a, double b);
+    public double sub(double a, double b);
+    public double add (double a, double b);
+    public double pow (double number, int power);
+    public double abs (double a);
+    public double sqrt (double a);
 }
