+++ HomeWork/src/HomeWork3/task8/CalculatorWithCounterDecorator.java
@@ -35,7 +35,7 @@ public class CalculatorWithCounterDecorator implements ICalculator {
      * @param b число на которое нужно разделить
      * @return результат деления числа a на b. Если b равно нулю вернет ноль.
      */
-    public float div(float a, float b){
+    public double div(double a, double b){
         counting();
         return this.calc.div(a,b);
     }
@@ -46,7 +46,7 @@ public class CalculatorWithCounterDecorator implements ICalculator {
      * @param b число на которое нужно умножить
      * @return результат умножения числа a на b.
      */
-    public float mult (float a, float b){
+    public double mult (double a, double b){
         counting();
         return this.calc.mult(a,b);
     }
@@ -57,7 +57,7 @@ public class CalculatorWithCounterDecorator implements ICalculator {
      * @param b число на которое нужно вычесть
      * @return результат вычитания числа a из числа b.
      */
-    public float sub(float a, float b){
+    public double sub(double a, double b){
         counting();
         return this.calc.sub(a,b);
     }
@@ -68,7 +68,7 @@ public class CalculatorWithCounterDecorator implements ICalculator {
      * @param b число которое нужно прибавить
      * @return результат сложения числа a и b.
      */
-    public float add (float a, float b){
+    public double add (double a, double b){
         counting();
         return this.calc.add(a,b);
     }
@@ -79,7 +79,7 @@ public class CalculatorWithCounterDecorator implements ICalculator {
      * @param power степень
      * @return результат возведения числа number d степень power.
      */
-    public float pow (float number, int power){
+    public double pow (double number, int power){
         counting();
         return this.calc.pow(number, power);
     }
@@ -89,7 +89,7 @@ public class CalculatorWithCounterDecorator implements ICalculator {
      * @param a число
      * @return модуль числа
      */
-    public float abs (float a){
+    public double abs (double a){
         counting();
         return this.calc.abs(a);
     }
@@ -99,7 +99,7 @@ public class CalculatorWithCounterDecorator implements ICalculator {
      * @param a число
      * @return результат извлечения квадратного корня числа. Если число отрицательное вернет ноль.
      */
-    public float sqrt (float a){
+    public double sqrt (double a){
         counting();
         return this.calc.sqrt(a);
     }
